+++ PixelGraph.Common/Material/MaterialSssProperties.cs
@@ -6,7 +6,7 @@ namespace PixelGraph.Common.Material
     {
         public ResourcePackSssChannelProperties Input {get; set;}
         public string Texture {get; set;}
-        public byte? Value {get; set;}
+        public decimal? Value {get; set;}
         public decimal? Scale {get; set;}
     }
 }
